--- test/raft_state_machine_test.cpp
@@ -1,675 +1,0 @@
-#include <gtest/gtest.h>
-#include <string>
-#include <memory>
-#include <filesystem>
-#include <atomic>
-#include <thread>
-#include <chrono>
-#include "butil/at_exit.h"
-#include "brpc/server.h"
-#include "braft/raft.h"
-#include "raft_state_machine.h"
-#include "http_server.h"
-#include "batched_indexer.h"
-#include "store.h"
-#include "threadpool.h"
-#include "http_data.h"
-#include "tsconfig.h"
-#include "collection_manager.h"
-#include "option.h"
-
-class ConfigImpl : public Config {
-public:
-    ConfigImpl(): Config() {}
-};
-
-class RaftStateMachineTest : public ::testing::Test {
-protected:
-    std::unique_ptr<butil::AtExitManager> exit_manager;
-    std::vector<std::unique_ptr<brpc::Server>> raft_servers;
-    Store* store;
-    Store* analytics_store;
-    ThreadPool* thread_pool;
-    ThreadPool* thread_pool_for_server;
-    HttpServer* http_server;
-    BatchedIndexer* batched_indexer;
-    http_message_dispatcher* message_dispatcher;
-    ConfigImpl* config;
-    std::string test_dir;
-    std::atomic<bool> quit;
-
-    void SetUp() override {
-        // Initialize braft dependencies first
-        exit_manager = std::make_unique<butil::AtExitManager>();
-
-        // Create test directory
-        test_dir = "/tmp/typesense_test/raft_state_machine";
-        std::filesystem::remove_all(test_dir);
-        std::filesystem::create_directories(test_dir);
-
-        // Initialize components
-        store = new Store(test_dir + "/store");
-        analytics_store = new Store(test_dir + "/analytics");
-
-        // Initialize CollectionManager - this is crucial for database operations
-        CollectionManager::get_instance().init(store, 1.0, "auth_key", quit);
-
-        thread_pool = new ThreadPool(4);
-
-        // Create minimal config for testing
-        config = new ConfigImpl();
-
-        // Create minimal real objects for testing
-        thread_pool_for_server = new ThreadPool(2);
-        http_server = new HttpServer("test", "127.0.0.1", 8080, "", "", 0, false, {}, thread_pool_for_server);
-        message_dispatcher = http_server->get_message_dispatcher();
-
-        // Initialize other components
-        batched_indexer = new BatchedIndexer(http_server, store, analytics_store, 4, *config, quit);
-
-        quit = false;
-    }
-
-    void TearDown() override {
-        delete batched_indexer;
-        thread_pool->shutdown();
-        delete thread_pool;
-        thread_pool_for_server->shutdown();
-        delete thread_pool_for_server;
-        delete store;
-        delete analytics_store;
-        delete config;
-        delete http_server;
-
-        // Stop all RPC servers
-        for (auto& server : raft_servers) {
-            if (server) {
-                server->Stop(0);
-                server->Join();
-            }
-        }
-        raft_servers.clear();
-
-        // Clean up test directory
-        std::filesystem::remove_all(test_dir);
-    }
-
-    std::unique_ptr<RaftStateMachine> createRaftStateMachine() {
-        return std::make_unique<RaftStateMachine>(
-            http_server, batched_indexer, store, analytics_store,
-            thread_pool, message_dispatcher, false, config, 4, 1000
-        );
-    }
-
-    brpc::Server* createRaftServer(const butil::EndPoint& endpoint) {
-        auto server = std::make_unique<brpc::Server>();
-        brpc::Server* server_ptr = server.get();
-
-        int add_service_result = braft::add_service(server_ptr, endpoint);
-        EXPECT_EQ(add_service_result, 0);
-
-        int start_result = server->Start(endpoint, nullptr);
-        EXPECT_EQ(start_result, 0);
-
-        raft_servers.push_back(std::move(server));
-        return server_ptr;
-    }
-};
-
-TEST_F(RaftStateMachineTest, Constructor) {
-    auto raft_state_machine = createRaftStateMachine();
-
-    EXPECT_NE(raft_state_machine, nullptr);
-    EXPECT_EQ(raft_state_machine->get_store(), store);
-    EXPECT_EQ(raft_state_machine->get_config(), config);
-    EXPECT_EQ(raft_state_machine->get_batched_indexer(), batched_indexer);
-    EXPECT_EQ(raft_state_machine->get_message_dispatcher(), message_dispatcher);
-}
-
-TEST_F(RaftStateMachineTest, Start) {
-    auto raft_state_machine = createRaftStateMachine();
-
-    // Set up raft startup parameters
-    butil::EndPoint peering_endpoint;
-    int result = butil::str2endpoint("127.0.0.1", 9000, &peering_endpoint);
-    EXPECT_EQ(result, 0);
-
-    int api_port = 9001;
-    int election_timeout_ms = 1000;
-    int snapshot_max_byte_count_per_rpc = 128 * 1024;
-    std::string raft_dir = test_dir + "/raft_startup";
-    std::string nodes_config = "127.0.0.1:9000:9001";
-    std::atomic<bool> quit_abruptly{false};
-
-    // Create raft directory with proper structure
-    std::filesystem::create_directories(raft_dir + "/log");
-    std::filesystem::create_directories(raft_dir + "/raft_meta");
-    std::filesystem::create_directories(raft_dir + "/snapshot");
-
-    // Set up RPC server for this endpoint
-    createRaftServer(peering_endpoint);
-
-    // RaftStateMachine should start successfully
-    int start_result = raft_state_machine->start(peering_endpoint, api_port, election_timeout_ms,
-                                                 snapshot_max_byte_count_per_rpc, raft_dir,
-                                                 nodes_config, quit_abruptly);
-    EXPECT_EQ(start_result, 0);
-
-    // State machine should be alive and ready
-    EXPECT_TRUE(raft_state_machine->is_alive());
-
-    // Should have proper raft state
-    auto status = raft_state_machine->get_status();
-    EXPECT_TRUE(status.contains("state"));
-    std::string state = status["state"];
-    EXPECT_TRUE(state == "LEADER" || state == "FOLLOWER" || state == "CANDIDATE");
-
-    raft_state_machine->shutdown();
-    std::filesystem::remove_all(raft_dir);
-}
-
-TEST_F(RaftStateMachineTest, IsAlive) {
-    auto raft_state_machine = createRaftStateMachine();
-
-    butil::EndPoint peering_endpoint;
-    int result = butil::str2endpoint("127.0.0.1", 9002, &peering_endpoint);
-    EXPECT_EQ(result, 0);
-
-    std::string raft_dir = test_dir + "/raft_alive";
-    std::filesystem::create_directories(raft_dir + "/log");
-    std::filesystem::create_directories(raft_dir + "/raft_meta");
-    std::filesystem::create_directories(raft_dir + "/snapshot");
-    std::atomic<bool> quit_abruptly{false};
-
-    createRaftServer(peering_endpoint);
-
-    int start_result = raft_state_machine->start(peering_endpoint, 9003, 1000,
-                                                 128 * 1024, raft_dir,
-                                                 "127.0.0.1:9002:9003", quit_abruptly);
-    EXPECT_EQ(start_result, 0);
-
-    // Should be alive after starting
-    EXPECT_TRUE(raft_state_machine->is_alive());
-
-    raft_state_machine->shutdown();
-    std::filesystem::remove_all(raft_dir);
-}
-
-TEST_F(RaftStateMachineTest, IsAliveWithoutStarting) {
-    auto raft_state_machine = createRaftStateMachine();
-
-    // Initially not alive
-    EXPECT_FALSE(raft_state_machine->is_alive());
-}
-
-TEST_F(RaftStateMachineTest, IsLeader) {
-    auto raft_state_machine = createRaftStateMachine();
-
-    butil::EndPoint peering_endpoint;
-    int result = butil::str2endpoint("127.0.0.1", 9004, &peering_endpoint);
-    EXPECT_EQ(result, 0);
-
-    std::string raft_dir = test_dir + "/raft_leader";
-    std::filesystem::create_directories(raft_dir + "/log");
-    std::filesystem::create_directories(raft_dir + "/raft_meta");
-    std::filesystem::create_directories(raft_dir + "/snapshot");
-    std::atomic<bool> quit_abruptly{false};
-
-    createRaftServer(peering_endpoint);
-
-    int start_result = raft_state_machine->start(peering_endpoint, 9005, 800,
-                                                 128 * 1024, raft_dir,
-                                                 "127.0.0.1:9004:9005", quit_abruptly);
-    EXPECT_EQ(start_result, 0);
-
-    // Wait for leader election to complete
-    std::this_thread::sleep_for(std::chrono::milliseconds(2500));
-
-    // Single node should elect itself as leader
-    EXPECT_TRUE(raft_state_machine->is_leader());
-
-    auto status = raft_state_machine->get_status();
-    EXPECT_EQ(status["state"], "LEADER");
-
-    raft_state_machine->shutdown();
-    std::filesystem::remove_all(raft_dir);
-}
-
-TEST_F(RaftStateMachineTest, IsLeaderWithoutStarting) {
-    auto raft_state_machine = createRaftStateMachine();
-
-    // Initially not leader
-    EXPECT_FALSE(raft_state_machine->is_leader());
-}
-
-TEST_F(RaftStateMachineTest, ReadWriteCaughtUp) {
-    auto raft_state_machine = createRaftStateMachine();
-
-    butil::EndPoint peering_endpoint;
-    int result = butil::str2endpoint("127.0.0.1", 9006, &peering_endpoint);
-    EXPECT_EQ(result, 0);
-
-    std::string raft_dir = test_dir + "/raft_ready";
-    std::filesystem::create_directories(raft_dir + "/log");
-    std::filesystem::create_directories(raft_dir + "/raft_meta");
-    std::filesystem::create_directories(raft_dir + "/snapshot");
-    std::atomic<bool> quit_abruptly{false};
-
-    createRaftServer(peering_endpoint);
-
-    int start_result = raft_state_machine->start(peering_endpoint, 9007, 1000,
-                                                 128 * 1024, raft_dir,
-                                                 "127.0.0.1:9006:9007", quit_abruptly);
-    EXPECT_EQ(start_result, 0);
-
-    // Wait for leader election
-    std::this_thread::sleep_for(std::chrono::milliseconds(2000));
-
-    // Should be ready for operations
-    EXPECT_TRUE(raft_state_machine->is_read_caught_up());
-    EXPECT_TRUE(raft_state_machine->is_write_caught_up());
-
-    raft_state_machine->shutdown();
-    std::filesystem::remove_all(raft_dir);
-}
-
-TEST_F(RaftStateMachineTest, ReadWriteCaughtUpWithoutStarting) {
-    auto raft_state_machine = createRaftStateMachine();
-
-    // Initially not ready
-    EXPECT_FALSE(raft_state_machine->is_read_caught_up());
-    EXPECT_FALSE(raft_state_machine->is_write_caught_up());
-}
-
-TEST_F(RaftStateMachineTest, GetStatus) {
-    auto raft_state_machine = createRaftStateMachine();
-
-    butil::EndPoint peering_endpoint;
-    int result = butil::str2endpoint("127.0.0.1", 9008, &peering_endpoint);
-    EXPECT_EQ(result, 0);
-
-    std::string raft_dir = test_dir + "/raft_status";
-    std::filesystem::create_directories(raft_dir + "/log");
-    std::filesystem::create_directories(raft_dir + "/raft_meta");
-    std::filesystem::create_directories(raft_dir + "/snapshot");
-    std::atomic<bool> quit_abruptly{false};
-
-    createRaftServer(peering_endpoint);
-
-    int start_result = raft_state_machine->start(peering_endpoint, 9009, 1000,
-                                                 128 * 1024, raft_dir,
-                                                 "127.0.0.1:9008:9009", quit_abruptly);
-    EXPECT_EQ(start_result, 0);
-
-    // Wait for startup to complete
-    std::this_thread::sleep_for(std::chrono::milliseconds(1500));
-
-    // Should be able to get status
-    auto status = raft_state_machine->get_status();
-    EXPECT_TRUE(status.contains("state"));
-
-    raft_state_machine->shutdown();
-    std::filesystem::remove_all(raft_dir);
-}
-
-TEST_F(RaftStateMachineTest, GetStatusWithoutStarting) {
-    auto raft_state_machine = createRaftStateMachine();
-
-    // Get status JSON
-    auto status = raft_state_machine->get_status();
-
-    // Should contain expected keys
-    EXPECT_TRUE(status.contains("state"));
-    EXPECT_TRUE(status.contains("committed_index"));
-    EXPECT_TRUE(status.contains("queued_writes"));
-    EXPECT_TRUE(status.contains("is_leader"));
-    EXPECT_TRUE(status.contains("read_ready"));
-    EXPECT_TRUE(status.contains("write_ready"));
-
-    // Initial values
-    EXPECT_EQ(status["state"], "NOT_READY");
-    EXPECT_EQ(status["committed_index"], 0);
-    EXPECT_EQ(status["is_leader"], false);
-    EXPECT_EQ(status["read_ready"], false);
-    EXPECT_EQ(status["write_ready"], false);
-}
-
-TEST_F(RaftStateMachineTest, GetLeaderUrl) {
-    auto raft_state_machine = createRaftStateMachine();
-
-    butil::EndPoint peering_endpoint;
-    int result = butil::str2endpoint("127.0.0.1", 9010, &peering_endpoint);
-    EXPECT_EQ(result, 0);
-
-    std::string raft_dir = test_dir + "/raft_leader_url";
-    std::filesystem::create_directories(raft_dir + "/log");
-    std::filesystem::create_directories(raft_dir + "/raft_meta");
-    std::filesystem::create_directories(raft_dir + "/snapshot");
-    std::atomic<bool> quit_abruptly{false};
-
-    createRaftServer(peering_endpoint);
-
-    int start_result = raft_state_machine->start(peering_endpoint, 9011, 1000,
-                                                 128 * 1024, raft_dir,
-                                                 "127.0.0.1:9010:9011", quit_abruptly);
-    EXPECT_EQ(start_result, 0);
-
-    // Wait for leader election
-    std::this_thread::sleep_for(std::chrono::milliseconds(2000));
-
-    // Should have leader URL
-    auto leader_url = raft_state_machine->get_leader_url();
-    EXPECT_FALSE(leader_url.empty());
-
-    raft_state_machine->shutdown();
-    std::filesystem::remove_all(raft_dir);
-}
-
-TEST_F(RaftStateMachineTest, GetLeaderUrlWithoutStarting) {
-    auto raft_state_machine = createRaftStateMachine();
-
-    // No leader initially
-    auto leader_url = raft_state_machine->get_leader_url();
-    EXPECT_TRUE(leader_url.empty());
-}
-
-TEST_F(RaftStateMachineTest, HasLeaderTerm) {
-    auto raft_state_machine = createRaftStateMachine();
-
-    butil::EndPoint peering_endpoint;
-    int result = butil::str2endpoint("127.0.0.1", 9012, &peering_endpoint);
-    EXPECT_EQ(result, 0);
-
-    std::string raft_dir = test_dir + "/raft_leader_term";
-    std::filesystem::create_directories(raft_dir + "/log");
-    std::filesystem::create_directories(raft_dir + "/raft_meta");
-    std::filesystem::create_directories(raft_dir + "/snapshot");
-    std::atomic<bool> quit_abruptly{false};
-
-    createRaftServer(peering_endpoint);
-
-    int start_result = raft_state_machine->start(peering_endpoint, 9013, 1000,
-                                                 128 * 1024, raft_dir,
-                                                 "127.0.0.1:9012:9013", quit_abruptly);
-    EXPECT_EQ(start_result, 0);
-
-    // Wait for leader election
-    std::this_thread::sleep_for(std::chrono::milliseconds(2000));
-
-    // Should have leader term
-    EXPECT_TRUE(raft_state_machine->has_leader_term());
-
-    raft_state_machine->shutdown();
-    std::filesystem::remove_all(raft_dir);
-}
-
-TEST_F(RaftStateMachineTest, HasLeaderTermWithoutStarting) {
-    auto raft_state_machine = createRaftStateMachine();
-
-    // No leader initially
-    bool has_leader = raft_state_machine->has_leader_term();
-    EXPECT_FALSE(has_leader);
-}
-
-TEST_F(RaftStateMachineTest, Write) {
-    auto raft_state_machine = createRaftStateMachine();
-
-    butil::EndPoint peering_endpoint;
-    int result = butil::str2endpoint("127.0.0.1", 9014, &peering_endpoint);
-    EXPECT_EQ(result, 0);
-
-    std::string raft_dir = test_dir + "/raft_write";
-    std::filesystem::create_directories(raft_dir + "/log");
-    std::filesystem::create_directories(raft_dir + "/raft_meta");
-    std::filesystem::create_directories(raft_dir + "/snapshot");
-    std::atomic<bool> quit_abruptly{false};
-
-    createRaftServer(peering_endpoint);
-
-    int start_result = raft_state_machine->start(peering_endpoint, 9015, 1000,
-                                                 128 * 1024, raft_dir,
-                                                 "127.0.0.1:9014:9015", quit_abruptly);
-    EXPECT_EQ(start_result, 0);
-
-    // Wait for leader election
-    std::this_thread::sleep_for(std::chrono::milliseconds(2000));
-
-    // Node should become leader
-    EXPECT_TRUE(raft_state_machine->is_leader());
-
-    // Should be ready for operations
-    EXPECT_TRUE(raft_state_machine->is_read_caught_up());
-    EXPECT_TRUE(raft_state_machine->is_write_caught_up());
-
-    // Test write request processing
-    auto request = std::make_shared<http_req>();
-    static int dummy_generator = 0;
-    auto response = std::make_shared<http_res>(&dummy_generator);
-
-    // Set up request for write
-    request->body = "{\"action\": \"create\", \"collection\": \"test\", \"id\": \"1\"}";
-    request->path_without_query = "/collections/test/documents";
-    request->route_hash = 12345;
-    request->http_method = "POST";
-    response->final = false;
-
-    // Write request should be processed
-    raft_state_machine->write(request, response);
-
-    // Give time for async processing
-    std::this_thread::sleep_for(std::chrono::milliseconds(500));
-
-    raft_state_machine->shutdown();
-    std::filesystem::remove_all(raft_dir);
-}
-
-TEST_F(RaftStateMachineTest, WriteWithoutStarting) {
-    auto raft_state_machine = createRaftStateMachine();
-
-    // Create test request and response
-    auto request = std::make_shared<http_req>();
-    static int dummy_generator = 0;
-    auto response = std::make_shared<http_res>(&dummy_generator);
-
-    // Set up request properties that write() method expects
-    request->body = "{\"test\": \"data\"}";
-    request->path_without_query = "/test";
-    request->route_hash = 12345;
-    response->final = false;
-
-    // This should not crash even without a started node
-    raft_state_machine->write(request, response);
-
-    // Response should have error status code (since no leader found)
-    EXPECT_EQ(response->status_code, 500);
-    EXPECT_NE(response->body.find("Could not find a leader"), std::string::npos);
-}
-
-TEST_F(RaftStateMachineTest, InitDb) {
-    auto raft_state_machine = createRaftStateMachine();
-
-    // Database initialization should succeed
-    int init_result = raft_state_machine->init_db();
-    EXPECT_EQ(init_result, 0);
-
-    // State machine should remain functional after init_db
-    EXPECT_NE(raft_state_machine, nullptr);
-
-    // Should be able to get status after initialization
-    auto status = raft_state_machine->get_status();
-    EXPECT_TRUE(status.contains("state"));
-
-    // Other operations should work correctly
-    EXPECT_FALSE(raft_state_machine->is_alive());
-}
-
-TEST_F(RaftStateMachineTest, TriggerVote) {
-    auto raft_state_machine = createRaftStateMachine();
-
-    butil::EndPoint peering_endpoint;
-    int result = butil::str2endpoint("127.0.0.1", 9016, &peering_endpoint);
-    EXPECT_EQ(result, 0);
-
-    std::string raft_dir = test_dir + "/raft_vote";
-    std::filesystem::create_directories(raft_dir + "/log");
-    std::filesystem::create_directories(raft_dir + "/raft_meta");
-    std::filesystem::create_directories(raft_dir + "/snapshot");
-    std::atomic<bool> quit_abruptly{false};
-
-    createRaftServer(peering_endpoint);
-
-    int start_result = raft_state_machine->start(peering_endpoint, 9017, 800,
-                                                 128 * 1024, raft_dir,
-                                                 "127.0.0.1:9016:9017", quit_abruptly);
-    EXPECT_EQ(start_result, 0);
-
-    // Wait for leader election
-    std::this_thread::sleep_for(std::chrono::milliseconds(2500));
-
-    // Should be able to trigger vote (no-op for leader)
-    raft_state_machine->trigger_vote();
-    // Result depends on node state
-
-    raft_state_machine->shutdown();
-    std::filesystem::remove_all(raft_dir);
-}
-
-TEST_F(RaftStateMachineTest, TriggerVoteWithoutStarting) {
-    auto raft_state_machine = createRaftStateMachine();
-
-    // Should fail without initialized node
-    bool vote_result = raft_state_machine->trigger_vote();
-    EXPECT_FALSE(vote_result);
-}
-
-TEST_F(RaftStateMachineTest, ResetPeers) {
-    auto raft_state_machine = createRaftStateMachine();
-
-    butil::EndPoint peering_endpoint;
-    int result = butil::str2endpoint("127.0.0.1", 9018, &peering_endpoint);
-    EXPECT_EQ(result, 0);
-
-    std::string raft_dir = test_dir + "/raft_reset";
-    std::filesystem::create_directories(raft_dir + "/log");
-    std::filesystem::create_directories(raft_dir + "/raft_meta");
-    std::filesystem::create_directories(raft_dir + "/snapshot");
-    std::atomic<bool> quit_abruptly{false};
-
-    createRaftServer(peering_endpoint);
-
-    int start_result = raft_state_machine->start(peering_endpoint, 9019, 1000,
-                                                 128 * 1024, raft_dir,
-                                                 "127.0.0.1:9018:9019", quit_abruptly);
-    EXPECT_EQ(start_result, 0);
-
-    // Wait for node to be ready
-    std::this_thread::sleep_for(std::chrono::milliseconds(2000));
-
-    // reset_peers should work
-    raft_state_machine->reset_peers();
-    // Result depends on implementation
-
-    raft_state_machine->shutdown();
-    std::filesystem::remove_all(raft_dir);
-}
-
-TEST_F(RaftStateMachineTest, ResetPeersWithoutStarting) {
-    auto raft_state_machine = createRaftStateMachine();
-
-    // Should fail without initialized node
-    bool reset_result = raft_state_machine->reset_peers();
-    EXPECT_FALSE(reset_result);
-}
-
-TEST_F(RaftStateMachineTest, RefreshNodes) {
-    auto raft_state_machine = createRaftStateMachine();
-
-    butil::EndPoint peering_endpoint;
-    int result = butil::str2endpoint("127.0.0.1", 9020, &peering_endpoint);
-    EXPECT_EQ(result, 0);
-
-    std::string raft_dir = test_dir + "/raft_refresh";
-    std::filesystem::create_directories(raft_dir + "/log");
-    std::filesystem::create_directories(raft_dir + "/raft_meta");
-    std::filesystem::create_directories(raft_dir + "/snapshot");
-    std::atomic<bool> quit_abruptly{false};
-
-    createRaftServer(peering_endpoint);
-
-    int start_result = raft_state_machine->start(peering_endpoint, 9021, 1000,
-                                                 128 * 1024, raft_dir,
-                                                 "127.0.0.1:9020:9021", quit_abruptly);
-    EXPECT_EQ(start_result, 0);
-
-    // Wait for cluster to be ready
-    std::this_thread::sleep_for(std::chrono::milliseconds(2000));
-
-    // Test refresh_nodes operations
-    std::string nodes_config = "127.0.0.1:9020:9021,127.0.0.1:9022:9023";
-    std::atomic<bool> reset_peers{false};
-
-    // refresh_nodes should work
-    raft_state_machine->refresh_nodes(nodes_config, 0, reset_peers);
-
-    // refresh_nodes with reset_peers should work
-    reset_peers = true;
-    raft_state_machine->refresh_nodes(nodes_config, 0, reset_peers);
-
-    // Node should remain functional after membership changes
-    auto status = raft_state_machine->get_status();
-    EXPECT_TRUE(status.contains("state"));
-
-    raft_state_machine->shutdown();
-    std::filesystem::remove_all(raft_dir);
-}
-
-TEST_F(RaftStateMachineTest, RefreshNodesWithoutStarting) {
-    auto raft_state_machine = createRaftStateMachine();
-
-    // Test node refresh operations
-    std::string nodes_config = "127.0.0.1:8090:8091";
-    std::atomic<bool> reset_peers{false};
-
-    // This should not crash even without initialized node
-    raft_state_machine->refresh_nodes(nodes_config, 0, reset_peers);
-}
-
-TEST_F(RaftStateMachineTest, RefreshCatchupStatus) {
-    auto raft_state_machine = createRaftStateMachine();
-
-    butil::EndPoint peering_endpoint;
-    int result = butil::str2endpoint("127.0.0.1", 9024, &peering_endpoint);
-    EXPECT_EQ(result, 0);
-
-    std::string raft_dir = test_dir + "/raft_catchup";
-    std::filesystem::create_directories(raft_dir + "/log");
-    std::filesystem::create_directories(raft_dir + "/raft_meta");
-    std::filesystem::create_directories(raft_dir + "/snapshot");
-    std::atomic<bool> quit_abruptly{false};
-
-    createRaftServer(peering_endpoint);
-
-    int start_result = raft_state_machine->start(peering_endpoint, 9025, 1000,
-                                                 128 * 1024, raft_dir,
-                                                 "127.0.0.1:9024:9025", quit_abruptly);
-    EXPECT_EQ(start_result, 0);
-
-    // Test refresh methods (should not crash)
-    raft_state_machine->refresh_catchup_status(false);
-    raft_state_machine->refresh_catchup_status(true);
-
-    raft_state_machine->shutdown();
-    std::filesystem::remove_all(raft_dir);
-}
-
-TEST_F(RaftStateMachineTest, RefreshCatchupStatusWithoutStarting) {
-    auto raft_state_machine = createRaftStateMachine();
-
-    // Should not crash when calling refresh_catchup_status without starting
-    raft_state_machine->refresh_catchup_status(true);
-    raft_state_machine->refresh_catchup_status(false);
-
-    // Should return false for readiness flags without starting
-    EXPECT_FALSE(raft_state_machine->is_read_caught_up());
-    EXPECT_FALSE(raft_state_machine->is_write_caught_up());
-}
